requirements.txt
azure-containerregistry==1.2.0
azure-identity==1.24.0
azure-mgmt-containerregistry==14.0.0
click==8.1.7
pexpect==4.9.0
PyYAML==6.0.1
